The local datastore was missing two rather critical constraints that are present in production: 1) Just as you can't perform writes involving entities in multiple entity groups within a single transaction, you also can't perform reads involving entities in multiple entity groups within a single transaction.
2) You can't write to the same entity multiple times within a single
transaction.

Sadly, the ORM was doing both of these things all over the place.  This CL
contains modifications that allow the ORM's testsuite to pass when running
against a local datastore that has these constraints.

The big design change is that 1-to-1 relationships no longer store the key of
the child object in the parent.  There is no way to make this work without
either a getNextId() function in the datastore or support for multiple writes to
a single entity within a transaction.  So, 1-to-1 relationships are implemented
the same way 1-to-many relationships are implemented: the parent has to query
the child table to find its children.  The difference here is that if the query
returns anything other than 0 or 1 results, it's a data error.

Another sizable design change:
We do not support non-owned relationships when running with a transactional
datasource.  This drastically simplifies the logic for determining whether or
not we should auto-assign entity group.  The answer is always yes.

I also updated the distro script to write MANIFEST.MF into the src tree.  Turned
out it was a mistake to have deleted it in an earlier CL.  This way we only need
to maintain the version number in one place.

Let's never do this again.

+++ src/org/datanucleus/store/appengine/DatastoreFKListStoreSpecialization.java
@@ -25,8 +25,6 @@ import org.datanucleus.util.Localiser;
 class DatastoreFKListStoreSpecialization extends DatastoreAbstractListStoreSpecialization
     implements FKListStoreSpecialization {
 
-  private static final int[] NOT_USED = {0};
-
   DatastoreFKListStoreSpecialization(Localiser localiser, ClassLoaderResolver clr,
       DatastoreManager storeMgr) {
     super(localiser, clr, storeMgr);
@@ -46,15 +44,20 @@ class DatastoreFKListStoreSpecialization extends DatastoreAbstractListStoreSpeci
     if (orderMapping != null) {
       ObjectManager om = ownerSm.getObjectManager();
       StateManager childSm = om.findStateManager(element);
-      Entity childEntity = handler.getAssociatedEntityForCurrentTransaction(childSm);
-      orderMapping.setObject(om, childEntity, NOT_USED, index);
-      handler.put(om, childEntity);
+      // See DatastoreFieldManager.handleIndexFields for info on why this
+      // absurdity is necessary.
+      Entity childEntity =
+          (Entity) childSm.getAssociatedValue(DatastorePersistenceHandler.ENTITY_WRITE_DELAYED);
+      if (childEntity != null) {
+        childSm.setAssociatedValue(DatastorePersistenceHandler.ENTITY_WRITE_DELAYED, null);
+        childSm.setAssociatedValue(ecs.getOrderMapping(), index);
+        handler.insertObject(childSm);
+      }
     }
 
-    if (ecs.getOwnerMemberMetaData().getCollection().isDependentElement() && allowDependentField) {
-      if (obj != null) {
-        ownerSm.getObjectManager().deleteObjectInternal(obj);
-      }
+    if (ecs.getOwnerMemberMetaData().getCollection().isDependentElement() &&
+        allowDependentField && obj != null) {
+      ownerSm.getObjectManager().deleteObjectInternal(obj);
     }
     return obj;
   }
@@ -67,13 +70,15 @@ class DatastoreFKListStoreSpecialization extends DatastoreAbstractListStoreSpeci
     StateManager elementSm = om.findStateManager(element);
     // The fk is already set but we still need to set the index
     DatastorePersistenceHandler handler = storeMgr.getPersistenceHandler();
-    Entity entity = handler.getAssociatedEntityForCurrentTransaction(elementSm);
-    if (entity == null) {
-      elementSm.locate();
+    // See DatastoreFieldManager.handleIndexFields for info on why this
+    // absurdity is necessary.
+    Entity entity =
+        (Entity) elementSm.getAssociatedValue(DatastorePersistenceHandler.ENTITY_WRITE_DELAYED);
+    if (entity != null) {
+      elementSm.setAssociatedValue(DatastorePersistenceHandler.ENTITY_WRITE_DELAYED, null);
+      elementSm.setAssociatedValue(ecs.getOrderMapping(), index);
+      handler.insertObject(elementSm);
     }
-    entity = handler.getAssociatedEntityForCurrentTransaction(elementSm);
-    ecs.getOrderMapping().setObject(om, entity, new int[1], index);
-    handler.put(om, entity);
     return true;
   }
 
@@ -92,9 +97,9 @@ class DatastoreFKListStoreSpecialization extends DatastoreAbstractListStoreSpeci
       // first we need to delete the element
       ObjectManager om = sm.getObjectManager();
       Object element = fkListStore.get(sm, index);
-      Key keyToRemove = this.extractElementKey(om, element);
+      StateManager elementStateManager = om.findStateManager(element);
       DatastorePersistenceHandler handler = storeMgr.getPersistenceHandler();
-      handler.delete(om, keyToRemove);
+      handler.deleteObject(elementStateManager);
       if (fkListStore.getOrderMapping() != null) {
         // now, if there is an order mapping, we need to shift
         // everyone down
